+++ core_packages/v_chat_sdk_core/lib/src/events/events.dart
@@ -1,7 +1,7 @@
 export './app_life_cycle.dart';
 export './socket_status_event.dart';
+export './v_call_events.dart';
 export './v_message_events.dart';
 export './v_notifications_events.dart';
 export './v_room_events.dart';
 export './v_socket_interval.dart';
-export './v_call_events.dart';