+++ core_packages/v_chat_sdk_core/lib/src/http/api_service/auth/auth_api_service.dart
@@ -3,8 +3,6 @@ import 'package:v_chat_sdk_core/src/http/api_service/interceptors.dart';
 import 'package:v_chat_sdk_core/v_chat_sdk_core.dart';
 import 'package:v_chat_utils/v_chat_utils.dart';
 
-import '../../../user_apis/auth/auth.dart';
-
 class VAuthApiService {
   VAuthApiService._();