+++ core_packages/v_chat_sdk_core/lib/src/http/api_service/call/call_api.dart
@@ -1,7 +1,6 @@
 import 'package:chopper/chopper.dart';
 import 'package:v_chat_sdk_core/src/http/api_service/interceptors.dart';
 import 'package:v_chat_sdk_core/src/utils/api_constants.dart';
-import 'package:v_chat_utils/v_chat_utils.dart';
 
 part 'call_api.chopper.dart';