+++ core_packages/v_chat_sdk_core/lib/src/http/socket/socket_service.dart
@@ -2,13 +2,10 @@ import 'dart:async';
 
 import 'package:logging/logging.dart';
 import 'package:v_chat_sdk_core/src/http/socket/socket_io_client.dart';
-import 'package:v_chat_sdk_core/src/models/socket/new_call_model.dart';
 import 'package:v_chat_sdk_core/src/native_api/local/native_local_message.dart';
 import 'package:v_chat_sdk_core/src/native_api/local/native_local_room.dart';
 import 'package:v_chat_sdk_core/v_chat_sdk_core.dart';
 import 'package:v_chat_utils/v_chat_utils.dart';
-
-import '../../models/socket/on_aceept_call.dart';
 
 class SocketService {
   final _log = Logger('SocketService');
@@ -143,7 +140,7 @@ class SocketService {
     _emitter.fire(VCallAcceptedEvent(
       roomId: onAcceptCall.roomId,
       data: onAcceptCall,
-    ));
+    ),);
   }
 
   void handleCallEnded(String roomId) {