+++ core_packages/v_chat_sdk_core/lib/src/models/controller/v_navigator.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/cupertino.dart';
-import 'package:v_chat_sdk_core/src/events/events.dart';
 import 'package:v_chat_sdk_core/src/models/models.dart';
 import 'package:v_chat_utils/v_chat_utils.dart';