+++ core_packages/v_chat_sdk_core/lib/src/service/call_listener.dart
@@ -1,7 +1,6 @@
 import 'package:logging/logging.dart';
+import 'package:v_chat_sdk_core/v_chat_sdk_core.dart';
 import 'package:v_chat_utils/v_chat_utils.dart';
-
-import '../../v_chat_sdk_core.dart';
 
 class CallListener {
   final VNativeApi nativeApi;