+++ v_ui_packages/v_chat_message_page/lib/src/page/message_page/v_message_page.dart
@@ -64,7 +64,7 @@ class _VMessagePageState extends State<VMessagePage> {
             return VMessageAppBare(
               state: value,
               onSearch: controller.onOpenSearch,
-              onCreateCall:  controller.onCreateCall,
+              onCreateCall: controller.onCreateCall,
               onViewMedia: () => controller.onViewMedia(context, value.roomId),
               onTitlePress: controller.onTitlePress,
             );