+++ v_ui_packages/v_chat_message_page/lib/src/widgets/app_bare/v_message_app_bare.dart
@@ -145,8 +145,6 @@ class VMessageAppBare extends StatelessWidget {
               PhosphorIcons.phoneCallFill,
             ),
           ),
-
-
         ],
       );
     }