+++ v_ui_packages/v_chat_web_rtc/lib/src/core/rtc_helper.dart
@@ -16,5 +16,4 @@ abstract class RtcHelper {
       {'DtlsSrtpKeyAgreement': true},
     ],
   };
-
 }