+++ v_ui_packages/v_chat_web_rtc/lib/src/pages/widgets/call_fotter.dart
@@ -11,7 +11,7 @@ class CallFotter extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return    Container(
+    return Container(
       padding: EdgeInsets.all(20),
       decoration: BoxDecoration(
         color: Color(0xff4f3434),