+++ v_ui_packages/v_chat_web_rtc/rtc_example/lib/app/modules/home/controllers/home_controller.dart
@@ -1,5 +1,3 @@
 import 'package:get/get.dart';
 
-class HomeController extends GetxController {
-
-}
+class HomeController extends GetxController {}